catalog/serializers.py
from rest_framework import serializers
from .models import Catalog

class CatalogSerializer(serializers.ModelSerializer):
    class Meta:
        model = Catalog
        fields = ['id', 'name', 'description', 'price', 'stock', 'image_url', 'category']
