requirements.txt
asgiref==3.9.2
asttokens==3.0.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
crispy-bootstrap4==2022.1
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
diff-match-patch==20241021
Django==5.2.7
django-allauth==65.12.0
django-crispy-forms==2.0
django-import-export==4.3.10
djangorestframework==3.16.1
exceptiongroup==1.3.0
executing==2.2.1
findspark==2.0.1
fonttools==4.60.0
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
matplotlib==3.7.1
matplotlib-inline==0.1.7
mrjob==0.7.4
nest-asyncio==1.6.0
numpy==1.24.3
packaging==25.0
pandas==1.5.3
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.1.0
pure_eval==0.2.3
py4j==0.10.9.7
Pygments==2.19.2
pyparsing==3.2.4
pyspark==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
pywin32==305
PyYAML==6.0.2
pyzmq==27.1.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.12.2
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tablib==3.8.0
threadpoolctl==3.6.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
wcwidth==0.2.13
